bugfix: [OPDS] Recreate a pdf chapter's cache if images are missing
When a user acccesses a PDF Chapter for the first time using the web reader, a corresponding cache directory is created containing only the pdf file. The pages are not extracted as png images into the cache directory.

If the user then requests a page from the same PDF Chapter using the OPDS controller/api, Kavita returns the error `No such image for page {0}. Try refreshing to allow re-cache.` .

This happens because the OPDS controller expects the chapter's pages (png images) to already exist in the cache directory. It does not attempt to extract them from the PDF file if the directory exists.

This code change fixes the issues by recreating the chapter's cache directory if chapter's cache directory exists and the opds contoller doesn't find in it the requested page by the user.

(I didn't look into how the fix behaves if the user tries to requests multiple pages in parallel)